backend/studybuddy_backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
firebase-admin
mysqlclient
Pillow
drf-yasg
coreapi